+++ src/ProductCatalog.Api/Services/IProductService.cs
@@ -7,8 +7,8 @@ namespace ProductCatalog.Api.Services;
 public interface IProductService
 {
     Task<PagedResult<Product>> SearchAsync(ProductQuery query, CancellationToken cancellationToken);
-    Task<Product> GetByIdAsync(long id, CancellationToken cancellationToken);
+    Task<Product> GetByIdAsync(int id, CancellationToken cancellationToken);
     Task<Product> CreateAsync(CreateProductRequest request, CancellationToken cancellationToken);
-    Task<Product> UpdateAsync(long id, UpdateProductRequest request, CancellationToken cancellationToken);
-    Task DeleteAsync(long id, CancellationToken cancellationToken);
+    Task<Product> UpdateAsync(int id, UpdateProductRequest request, CancellationToken cancellationToken);
+    Task DeleteAsync(int id, CancellationToken cancellationToken);
 }
